Project General Structure
You can use the "VeriSetiOlusturma" file to create a dataset  according to your input. There is variable features such as standard deviation of the height of apples,standard deviation of the weight of apples, standard deviation of the height of oranges,standard deviation of the weight of oranges.. For applying SVM algorithm you can use "SVM Algorithm" with a dataset and you can use the result in graph with using "Grafik Oluşturma" file. There are some example of dataset in the project.